退坑skyblock

+++ src/main/java/cn/seiua/skymatrix/hud/Hud.java
@@ -3,7 +3,7 @@ package cn.seiua.skymatrix.hud;
 import net.minecraft.client.util.math.MatrixStack;
 
 public interface Hud {
-    void draw(MatrixStack matrixStack, int x, int y);
+    void draw(MatrixStack matrixStack, float x, float y);
 
 
     int getHudWidth();
